mess-meal-planner-backend/requirements.txt
absl-py==2.3.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.5
attrs==25.3.0
av==14.4.0
babel==2.17.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.13.4
bidict==0.23.1
bleach==6.2.0
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
catboost==1.2.8
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.3.0
click==8.2.1
cloudpickle==3.1.1
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.2
ctranslate2==4.6.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
durationpy==0.10
easyocr==1.7.2
et_xmlfile==2.0.0
executing==2.1.0
Faker==33.1.0
fastapi==0.120.3
faster-whisper==1.1.1
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fonttools==4.58.1
fqdn==1.5.1
frozenlist==1.6.2
fsspec==2025.5.1
gevent==25.5.1
gevent-websocket==0.10.1
google-ai-generativelanguage==0.6.18
google-api-core==2.25.0
google-api-python-client==2.186.0
google-auth==2.40.3
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
graphviz==0.20.3
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imbalanced-learn==0.13.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.6.1
jaxlib==0.6.1
jedi==0.19.1
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
kubernetes==34.1.0
langchain==0.3.25
langchain-chroma==1.0.0
langchain-community==0.3.24
langchain-core==1.0.2
langchain-google-genai==2.1.5
langchain-text-splitters==0.3.8
langsmith==0.3.45
lazy_loader==0.4
lightgbm==4.6.0
lime==0.2.0.1
llvmlite==0.44.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
mistune==3.1.3
ml_dtypes==0.5.1
mmh3==5.2.0
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.42.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
notebook==7.4.3
notebook_shim==0.2.4
numba==0.61.2
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.22.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opt_einsum==3.4.0
orjson==3.10.18
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.6
plotly==6.1.2
posthog==5.4.0
prometheus_client==0.22.1
prompt_toolkit==3.0.48
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.33.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.18.0
PyMuPDF==1.26.5
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.2
python-json-logger==3.3.0
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
regex==2025.10.23
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.2.0
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.16.3
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==5.1.2
sentencepiece==0.2.0
setuptools==75.6.0
shap==0.47.2
shapely==2.1.2
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sklearn-compat==0.1.3
slicer==0.0.8
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.49.1
sympy==1.14.0
tabula==1.0.5
tabula-py==2.10.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.6.1
tinycss2==1.4.0
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typer==0.20.0
types-python-dateutil==2.9.0.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.34.3
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wget==3.2
widgetsnbextension==4.0.14
wrapt==1.17.0
wsproto==1.2.0
xgboost==3.0.2
yarl==1.20.0
zipp==3.23.0
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
